# swagger-ui/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "grpc-rest-go",
    "version": "1.0",
    "contact": {
      "name": "grpc-rest-go",
      "url": "https://github.com/romangurevitch/grpc-rest-go",
      "email": "none@example.com"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/dictionary/add/{word}": {
      "put": {
        "summary": "Add a word to the dictionary",
        "operationId": "AddWord",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiResult"
            }
          }
        },
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "DictionaryService"
        ]
      }
    },
    "/api/v1/dictionary/delete/{word}": {
      "delete": {
        "summary": "Delete a word from the dictionary",
        "operationId": "DeleteWord",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiResult"
            }
          }
        },
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "DictionaryService"
        ]
      }
    },
    "/api/v1/dictionary/find/{word}": {
      "get": {
        "summary": "Find a word in the dictionary",
        "operationId": "FindWord",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiFindWordRes"
            }
          }
        },
        "parameters": [
          {
            "name": "word",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "DictionaryService"
        ]
      }
    },
    "/api/v1/dictionary/popular": {
      "get": {
        "summary": "Get most popular words in the dictionary",
        "operationId": "GetMostPopularWords",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetMostPopularWordsRes"
            }
          }
        },
        "tags": [
          "DictionaryService"
        ]
      }
    }
  },
  "definitions": {
    "ResultStatus": {
      "type": "string",
      "enum": [
        "Success",
        "Fail"
      ],
      "default": "Success"
    },
    "apiFindWordRes": {
      "type": "object",
      "properties": {
        "exist": {
          "type": "boolean",
          "format": "boolean"
        },
        "result": {
          "$ref": "#/definitions/apiResult"
        }
      }
    },
    "apiGetMostPopularWordsRes": {
      "type": "object",
      "properties": {
        "words": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          }
        },
        "result": {
          "$ref": "#/definitions/apiResult"
        }
      }
    },
    "apiResult": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/ResultStatus"
        },
        "error": {
          "type": "string"
        }
      }
    }
  }
}